combined actions and reducers for retailers and organizations
