+++ src/actions/index.js
@@ -20,27 +20,20 @@ export {
   USER_SEARCH_START,
   USER_SEARCH_SUCCESS,
   USER_SEARCH_FAILURE,
-  userSearchResults
-} from "./searchUser";
-export {
-  ADD_SYSTEM_USER_START,
-  ADD_SYSTEM_USER_SUCCESS,
-  ADD_SYSTEM_USER_FAILURE,
-  addNewSystemUser
-} from "./addNewSystemUser";
-export {
+  userSearchResults,
   UPDATE_SYSTEM_USER_SUCCESS,
   UPDATE_SYSTEM_USER_START,
   UPDATE_SYSTEM_USER_FAILURE,
-  updateSystemUser
-} from "./updateSystemUser";
-export {
+  updateSystemUser,
+  ADD_SYSTEM_USER_START,
+  ADD_SYSTEM_USER_SUCCESS,
+  ADD_SYSTEM_USER_FAILURE,
+  addNewSystemUser,
   DELETE_SYSTEM_USER_START,
   DELETE_SYSTEM_USER_SUCCESS,
   DELETE_SYSTEM_USER_FAILURE,
   deleteSystemUser
-} from "./deleteSystemUser";
-// import { fromEventPattern } from "rxjs";
+} from "./userActions";
 
 export {
   DATA_TRANSACTION_CARD_START,
@@ -121,3 +114,25 @@ export {
   deleteFarmer,
   searchFarmers
 } from "./farmerAction";
+
+export {
+  RETAILER_SEARCH_FAILURE,
+  RETAILER_SEARCH_SUCCESS,
+  RETAILER_SEARCH_START,
+  searchRetailers,
+  ADD_RETAILER_START,
+  ADD_RETAILER_SUCCESS,
+  ADD_RETAILER_FAILURE,
+  addRetailer
+} from "./retailerActions"
+
+export {
+  ADD_ORGANIZATION_START,
+  ADD_ORGANIZATION_SUCCESS,
+  ADD_ORGANIZATION_FAILURE,
+  addOrganization,
+  ORGANIZATION_SEARCH_START,
+  ORGANIZATION_SEARCH_SUCCESS,
+  ORGANIZATION_SEARCH_FAILURE,
+  searchOrganizations
+} from "./"
