+++ src/reducers/index.js
@@ -1,13 +1,9 @@
 import userReducer from "../reducers/UserReducer";
-import reducerNewSystemUser from "../reducers/reducerNewSystemUser";
 import { combineReducers } from "redux";
 import loginReducer from "../reducers/loginReducer";
-import reducerProductStatsVue from "../reducers/reducerProductStatsVue";
-import retailerSearchReducer from "../reducers/retailerSearchReducer";
+import retailerReducer from "../reducers/retailerReducer";
 import reducerTransactionCard from "../reducers/reducerTransactionCard";
 import reducerInstallmentCard from "../reducers/reducerInstallmentCard";
-import reducerYieldCard from "../reducers/reducerYieldCard";
-import reducerInventoryCard from "../reducers/reducerInventoryCard";
 import farmerReducer from "../reducers/farmerReducer";
 import organizationSearchReducer from "../reducers/organizationSearchReducer";
 import inventoryReducer from "../reducers/inventoryReducer";
@@ -19,14 +15,10 @@ import itemTypeReducer from "./itemTypeReducer.js";
 
 export default combineReducers({
   userReducer: userReducer,
-  addSystemUserReducer: reducerNewSystemUser,
   login: loginReducer,
-  prodStatData: reducerProductStatsVue,
   transactionCardData: reducerTransactionCard,
   installmentCardData: reducerInstallmentCard,
-  inventoryCardData: reducerInventoryCard,
-  yieldCardData: reducerYieldCard,
-  retailerSearchData: retailerSearchReducer,
+  retailerData: retailerReducer,
   farmerData: farmerReducer,
   organizationSearchData: organizationSearchReducer,
   inventory: inventoryReducer,
